refactor: Separation of components and functions
- Change code that requires repetition to use useMemo hook
- Separate page-by-page components
