+++ package.json
@@ -14,8 +14,8 @@
     "@tanstack/react-query": "^5.40.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "embla-carousel-react": "^8.1.5",
-    "framer-motion": "^11.2.12",
+    "embla-carousel-react": "^8.1.6",
+    "framer-motion": "^11.2.13",
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.379.0",
     "next": "14.2.3",
